update result and error code, update employee service impl, update deparment service impl, add global exception handler, update r2 code

+++ src/main/java/io/clementleetimfu/educationmanagementsystem/service/impl/EmployeeServiceImpl.java
@@ -2,6 +2,8 @@ package io.clementleetimfu.educationmanagementsystem.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import io.clementleetimfu.educationmanagementsystem.exception.BusinessException;
+import io.clementleetimfu.educationmanagementsystem.exception.ErrorCodeEnum;
 import io.clementleetimfu.educationmanagementsystem.mapper.EmployeeMapper;
 import io.clementleetimfu.educationmanagementsystem.mapper.WorkExperienceMapper;
 import io.clementleetimfu.educationmanagementsystem.pojo.PageResult;
@@ -39,6 +41,11 @@ public class EmployeeServiceImpl implements EmployeeService {
 
         List<EmployeeSearchResponseDTO> employeeSearchResponseDTOList = employeeMapper.searchEmployee(employeeSearchRequestDTO);
 
+        if (employeeSearchResponseDTOList.isEmpty()) {
+            log.warn("Employee list is empty");
+            throw new BusinessException(ErrorCodeEnum.EMPLOYEE_NOT_FOUND);
+        }
+
         Page<EmployeeSearchResponseDTO> page = (Page<EmployeeSearchResponseDTO>) employeeSearchResponseDTOList;
         return new PageResult<>(page.getTotal(), page.getResult());
     }
@@ -50,9 +57,13 @@ public class EmployeeServiceImpl implements EmployeeService {
         employee.setCreateTime(LocalDateTime.now());
         employee.setUpdateTime(LocalDateTime.now());
         employee.setIsDeleted(false);
-        Integer addEmployeeResult = employeeMapper.addEmployee(employee);
+        Integer addEmployeeRowsAffected = employeeMapper.addEmployee(employee);
 
-        Integer addWorkExperienceResult = null;
+        if (addEmployeeRowsAffected == 0) {
+            log.warn("Failed to add employee:{}", employee);
+            throw new BusinessException(ErrorCodeEnum.EMPLOYEE_ADD_FAILED);
+        }
+
         List<WorkExperienceAddDTO> workExperienceAddDTOList = employeeAddDTO.getWorkExpList();
         if (!workExperienceAddDTOList.isEmpty()) {
             List<WorkExperience> workExperienceList = workExperienceAddDTOList.stream()
@@ -64,24 +75,45 @@ public class EmployeeServiceImpl implements EmployeeService {
                         workExperience.setIsDeleted(Boolean.FALSE);
                         return workExperience;
                     }).toList();
-            addWorkExperienceResult = workExperienceMapper.addWorkExperienceByBatch(workExperienceList);
-        } else {
-            addWorkExperienceResult = 1;
+            Integer addWorkExperienceRowsAffected = workExperienceMapper.addWorkExperienceByBatch(workExperienceList);
+
+            if (addWorkExperienceRowsAffected == 0) {
+                log.warn("Failed to add work experience:{}", employee);
+                throw new BusinessException(ErrorCodeEnum.WORK_EXPERIENCE_ADD_FAILED);
+            }
         }
 
-        return addEmployeeResult > 0 && addWorkExperienceResult > 0;
+        return Boolean.TRUE;
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
     public Boolean deleteEmployeeByIds(List<Integer> ids) {
-        Integer deleteEmployeeResult = employeeMapper.deleteEmployeeByIds(ids);
-        Integer deleteWorkExperienceResult = workExperienceMapper.deleteWorkExperienceByEmpIds(ids);
-        return deleteEmployeeResult > 0 && deleteWorkExperienceResult > 0;
+        Integer deleteEmployeeRowsAffected = employeeMapper.deleteEmployeeByIds(ids);
+
+        if (deleteEmployeeRowsAffected == 0) {
+            log.warn("Failed to delete employee with ids:{}", ids);
+            throw new BusinessException(ErrorCodeEnum.EMPLOYEE_DELETE_FAILED);
+        }
+
+        Integer deleteWorkExperienceRowsAffected = workExperienceMapper.deleteWorkExperienceByEmpIds(ids);
+
+        if (deleteWorkExperienceRowsAffected == 0) {
+            log.warn("Failed to delete work experience with employee ids:{}", ids);
+            throw new BusinessException(ErrorCodeEnum.WORK_EXPERIENCE_DELETE_FAILED);
+        }
+
+        return Boolean.TRUE;
     }
 
     @Override
     public EmployeeFindByIdDTO findEmployeeById(Integer id) {
-        return employeeMapper.findEmployeeById(id);
+        EmployeeFindByIdDTO employeeFindByIdDTO = employeeMapper.findEmployeeById(id);
+        if (employeeFindByIdDTO == null) {
+            log.warn("Employee with id:{} not found", id);
+            throw new BusinessException(ErrorCodeEnum.EMPLOYEE_NOT_FOUND);
+        }
+        return employeeFindByIdDTO;
     }
 
     @Transactional(rollbackFor = Exception.class)
@@ -89,10 +121,20 @@ public class EmployeeServiceImpl implements EmployeeService {
     public Boolean updateEmployee(EmployeeUpdateDTO employeeUpdateDTO) {
         Employee employee = modelMapper.map(employeeUpdateDTO, Employee.class);
         employee.setUpdateTime(LocalDateTime.now());
-        Integer updateEmployeeResult = employeeMapper.updateEmployee(employee);
-        Integer deleteWorkExperienceResult = workExperienceMapper.deleteWorkExperienceByEmpIds(Arrays.asList(employee.getId()));
+        Integer updateEmployeeRowsAffected = employeeMapper.updateEmployee(employee);
 
-        Integer updateWorkExperienceResult = null;
+        if (updateEmployeeRowsAffected == 0) {
+            log.warn("Failed to update employee:{}", employee);
+            throw new BusinessException(ErrorCodeEnum.EMPLOYEE_UPDATE_FAILED);
+        }
+
+        Integer deleteWorkExperienceRowsAffected = workExperienceMapper.deleteWorkExperienceByEmpIds(Arrays.asList(employee.getId()));
+
+        if (deleteWorkExperienceRowsAffected == 0) {
+            log.warn("Failed to delete work experience with employee id:{}", employee.getId());
+            throw new BusinessException(ErrorCodeEnum.WORK_EXPERIENCE_DELETE_FAILED);
+        }
+
         List<WorkExperienceUpdateDTO> workExperienceUpdateDTOList = employeeUpdateDTO.getWorkExpList();
         if (!workExperienceUpdateDTOList.isEmpty()) {
             List<WorkExperience> workExperienceList = workExperienceUpdateDTOList.stream().map(workExperienceUpdateDTO -> {
@@ -103,10 +145,15 @@ public class EmployeeServiceImpl implements EmployeeService {
                 workExperience.setIsDeleted(Boolean.FALSE);
                 return workExperience;
             }).toList();
-            updateWorkExperienceResult = workExperienceMapper.addWorkExperienceByBatch(workExperienceList);
-        } else {
-            updateWorkExperienceResult = 1;
+            Integer addWorkExperienceRowsAffected = workExperienceMapper.addWorkExperienceByBatch(workExperienceList);
+
+            if (addWorkExperienceRowsAffected == 0) {
+                log.warn("Failed to add work experience:{}", employee);
+                throw new BusinessException(ErrorCodeEnum.WORK_EXPERIENCE_ADD_FAILED);
+            }
+
         }
-        return updateEmployeeResult > 0 && deleteWorkExperienceResult > 0 && updateWorkExperienceResult > 0;
+
+        return Boolean.TRUE;
     }
 }
